moving where the conf file is being called

+++ snapshot2/snapshot_prod-dump.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 set -uex
-PRODDB="$1"
-DUMPDIR="$2/${PRODDB}"
+
+[ ! -f /etc/dbdump/conf ] && exit 1
+source /etc/dbdump/conf
+
+PRODDB="${1}"
+DUMPDIR="${PRODDUMPDIR}/${PRODDB}"
 [ ! -d "${DUMPDIR}/" ] && mkdir -p "${DUMPDIR}/"
 rm -f ${DUMPDIR}/*
 mysqldump ${PRODDB} --single-transaction --tab=${DUMPDIR}/
